Split RocketConfigs into separate file
